Merge installcd and admincd for amd64
This merges the two offerings together, as we are not limited to CDROM sizes in 2025.

There are also other quailty of life features such as using NetworkManager by default and just enough portage that user could use this new installcd as a way to recover a systemm using emerge --root.

Filesizes:

Installcd - 605MB
Admincd - 800MB
MergedCD - 734MB

Signed-off-by: Ian Jordan <immoloism@gmail.com>